Rename fs.c to sys.c to include more useful functions
System is more broad the fs so we can include other system related functions.
Also making file swupd indepentent. Only system specific functions should move
to this file
