+++ src/manifest.c
@@ -1300,3 +1300,18 @@ int enforce_compliant_manifest(struct file **a, struct file **b, int searchsize,
 	}
 	return ret; // If collisions were found, so manifest is purely additive
 }
+
+long get_manifest_list_contentsize(struct list *manifests)
+{
+	long total_size = 0;
+
+	struct list *ptr = NULL;
+	for (ptr = list_head(manifests); ptr; ptr = ptr->next) {
+		struct manifest *subman;
+		subman = ptr->data;
+
+		total_size += subman->contentsize;
+	}
+
+	return total_size;
+}
