+++ src/sys.h
@@ -1,0 +1,14 @@
+#ifndef __SWUPD_SYS__
+#define __SWUPD_SYS__
+
+#ifdef __cplusplus
+extern "C" {
+#endif
+
+extern long get_available_space(const char *path);
+
+#ifdef __cplusplus
+}
+#endif
+
+#endif
